Removing buffer allocations from benchmark methods.
